requirements.txt
Django
djangorestframework
mysqlclient
debugpy==1.5.1